Dev/putting logs in db (#8)
logging system was enhanced by updating IKLogTool and AbstractLogTool to support default attribute-based locations and passing location data to implementers. Logging was added to the DataGateway, along with summary information inside logs. A new MySQLLogger was introduced with console output, accompanied by tests and a sample test configuration. The ExpenseService was updated to use the new logger, and the .gitignore file was cleaned up.